Refactor imports to use .js extensions; add validation middleware for user credentials and signup

+++ server/controllers/invite.js
@@ -1,14 +1,14 @@
-import { prisma } from "../prisma/client";
+import { prisma } from "../prisma/client.js";
 
 export const create = async (req, res, next) => {
   const { recipientId } = req.body;
   try {
-    const recipient = await prisma.user.findUnique({ where: { id: recipientId }});
+    const recipient = await prisma.user.findUnique({ where: { id: parseInt(recipientId) }});
     if(!recipient) return res.status(200).json({ msg: "User does not exist" });
-    const invite = await prisma.invite.create({
+    await prisma.invite.create({
       data: {
         senderId: req.user.id,
-        recipientId,
+        recipientId: parseInt(recipientId),
       }
     });
     res.status(200).json({ msg: "Sent" });
@@ -20,10 +20,21 @@ export const create = async (req, res, next) => {
 export const received = async (req, res, next) => {
   try {
     const invites = await prisma.invite.findMany({
-      where: { recipientId: req.user.id },
-      include: { sender: true },
+      where: { 
+        recipientId: req.user.id
+      },
+      include: {
+        sender: {
+          select: {
+            id: true,
+            name: true,
+            avatarUrl: true,
+          }
+        }
+      },
     });
-    res.status(200).json({ invites });
+    const formattedInvites = invites.map(i => i.sender);
+    res.status(200).json({ invites: formattedInvites });
   } catch (error) {
     next(error);
   }
@@ -34,10 +45,17 @@ export const sent = async (req, res, next) => {
     const invites = await prisma.invite.findMany({
       where: { senderId: req.user.id },
       include: {
-        recipient: true
+        recipient: {
+           select: {
+            id: true,
+            name: true,
+            avatarUrl: true,
+          }
+        }
       }
     });
-    res.status(200).json({ invites });
+    const formattedInvites = invites.map(i => i.sender);
+    res.status(200).json({ invites: formattedInvites });
   } catch (error) {
     next(error);
   }
@@ -46,6 +64,13 @@ export const sent = async (req, res, next) => {
 export const accept = async (req, res, next) => {
   const { senderId } = req.body;
   try {
+    const invite = await prisma.invite.findUnique({
+       senderId_recipientId: {
+          senderId,
+          recipientId: req.user.id,
+        },
+    });
+    if(!invite) return res.status(400).json({ msg: 'Invite does not exist' })
     await prisma.invite.delete({
       where: {
         senderId_recipientId: {
@@ -55,13 +80,13 @@ export const accept = async (req, res, next) => {
       },
      });
     const chat = await prisma.chat.create();
-    const sender = await prisma.chatMember.create({
+    await prisma.chatMember.create({
       data: {
         userId: senderId,
         chatId: chat.id
       }
     });
-    const recipient = await prisma.chatMember.create({
+    await prisma.chatMember.create({
      data: {
         userId: req.user.id,
         chatId: chat.id
@@ -76,6 +101,13 @@ export const accept = async (req, res, next) => {
 export const decline = async (req, res, next) => {
   const { senderId } = req.body;
   try {
+    const invite = await prisma.invite.findUnique({
+       senderId_recipientId: {
+          senderId,
+          recipientId: req.user.id,
+        },
+    });
+    if(!invite) return res.status(400).json({ msg: 'Invite does not exist' })
     await prisma.invite.delete({
       where: {
         senderId_recipientId: {
@@ -95,6 +127,13 @@ export const decline = async (req, res, next) => {
 export const remove = async (req, res, next) => {
   const { recipientId } = req.body;
   try {
+    const invite = await prisma.invite.findUnique({
+       senderId_recipientId: {
+          senderId: req.user.id,
+          recipientId,
+        },
+    });
+    if(!invite) return res.status(400).json({ msg: 'Invite does not exist' })
     await prisma.invite.delete({
       where: {
         senderId_recipientId: {
